rename singnin.js to ApiManager.js, add comment explaining structure transformation to add-puzzle, perform general cleanup
